+++ nodejs/src/utils.js
@@ -11,31 +11,33 @@ if ( !String.prototype.isEmpty ) {
     };
 }
 
-const addTimestamp = winston.format( ( info ) => {
+const enrichLogContext = winston.format( ( info ) => {
     info['timestamp'] = new Date().toJSON();
+    info['worker_id'] = process.env.worker_id || 'primary';
     return info;
 } );
 
 const ini = require( "node-ini" );
 const config = ini.parseSync( path.resolve( __dirname, '../config.ini' ) );
-exports.logger = winston.createLogger( {
-    levels: winston.config.syslog.levels,
+const logger = winston.createLogger( {
+    levels: winston.config.npm.levels,
     transports: [
         new winston.transports.Console( {
             level: config.log.level,
-            format: format.combine( addTimestamp(), format.json() ),
+            format: format.combine( enrichLogContext(), format.json() ),
         } ),
         new winston.transports.DailyRotateFile( {
             filename: path.resolve( __dirname, config.log.file ),
             zippedArchive: true,
             maxSize: '100m',
             level: config.log.level,
-            format: format.combine( addTimestamp(), format.json() ),
+            format: format.combine( enrichLogContext(), format.json() ),
         } ),
     ],
 } );
+exports.logger = logger;
 
-let parseHeaderRemoteAddress = ( headersList ) => {
+const parseHeaderRemoteAddress = ( headersList ) => {
     const ipRegex = require( 'ip-regex' );
 
     for ( let key of [
@@ -60,3 +62,32 @@ exports.getWebSocketClientAddress = ( socket ) => {
     let remoteAddress = parseHeaderRemoteAddress( socket.handshake.headers );
     return remoteAddress ? remoteAddress : socket.handshake.address;
 };
+
+exports.setMonitoring = () => {
+
+    const v8 = require( 'v8' );
+    // enabling trace-gc
+    v8.setFlagsFromString( '--trace-gc' );
+    const {PerformanceObserver} = require( 'node:perf_hooks' );
+    // Create a performance observer
+    const obs = new PerformanceObserver( list => {
+        const entry = list.getEntries()[0];
+        /*
+        The entry is an instance of PerformanceEntry containing
+        metrics of a single garbage collection event.
+        For example,
+        PerformanceEntry {
+          name: 'gc',
+          entryType: 'gc',
+          startTime: 2820.567669,
+          duration: 1.315709,
+          kind: 1
+        }
+        */
+        logger.verbose( ['GC: ', entry] );
+    } );
+
+    obs.observe( {entryTypes: ['gc']} );
+    setInterval( () => { logger.verbose( ['Memory: ', process.memoryUsage()] ); }, 30000 );
+
+}
